Create Exercise 1: Greeting App
Develop a GUI to greet the user. The app should have two frames: InputFrame and DisplayFrame.

In InputFrame, include the following:

A title label in blue.
An entry field for the user's name.
A dropdown menu for selecting a color.
An "Update Greeting" button.
In DisplayFrame, include a label to display the personalized greeting.
Initially, DisplayFrame is empty, when the user clicks the "Update Greeting" button in InputFrame, the personalized greeting should appear in DisplayFrame with the selected color.
Use different background colors for each frame.